#24 [refactor] beforeEach code refactoring
